When player wants to change tiles
Added change-tiles button, the player can double-click to mark the tiles they want to change(or unmark to not change them anymore), the changed tiles get placed back into the bag and gets shuffled. The player cannot place a tile on the board AND change tiles. If they change their tiles then it's next turn